+++ ahcore/models/embedding_functions/vit_embed.py
@@ -1,42 +1,13 @@
-from abc import ABC, abstractmethod
-from enum import Enum
-from typing import Any, Callable
+from typing import Optional
 
 import torch
 
-from ahcore.utils.types import OutputModeBase, ViTEmbedMode
-
-
-class _TokenNames(str, Enum):
-    def __str__(self) -> Any:
-        return self.value
-
-
-class ViTTokenNames(_TokenNames):
-    CLS_TOKEN_NAME = "x_norm_clstoken"
-    PATCH_TOKEN_NAME = "x_norm_patchtokens"
-
-
-class AhCoreFeatureEmbedding(ABC):
-    def __init__(self, embedding_mode: OutputModeBase):
-        self._embedding_mode = embedding_mode
-        self.embed_fn: Callable[[dict[str, Any]], torch.Tensor]
-
-    @property
-    def embedding_mode(self) -> Any:
-        return self._embedding_mode.value
-
-    @property
-    def token_names(self) -> Any:
-        raise NotImplementedError
-
-    @abstractmethod
-    def _set_embed_function(self) -> None:
-        raise NotImplementedError
+from ahcore.models.embedding_functions.base_embedding import AhCoreFeatureEmbedding
+from ahcore.utils.types import ViTEmbedMode
 
 
 class ViTEmbed(AhCoreFeatureEmbedding):
-    def __init__(self, embedding_mode: OutputModeBase):
+    def __init__(self, embedding_mode: str):
         super().__init__(embedding_mode=embedding_mode)
         self._set_embed_function()
 
@@ -52,10 +23,6 @@ class ViTEmbed(AhCoreFeatureEmbedding):
         else:
             return 1
 
-    @property
-    def token_names(self) -> tuple[ViTTokenNames, ViTTokenNames]:
-        return ViTTokenNames.CLS_TOKEN_NAME, ViTTokenNames.PATCH_TOKEN_NAME
-
     def _set_embed_function(self) -> None:
         if self._embedding_mode == ViTEmbedMode.CLS_ONLY:
             self.embed_fn = self.embed_cls_only
@@ -70,37 +37,59 @@ class ViTEmbed(AhCoreFeatureEmbedding):
         else:
             raise NotImplementedError(f"Embedding mode {self._embedding_mode} is not supported.")
 
-    @staticmethod
-    def get_output_tokens(jit_model_prediction_sample: dict[str, Any]) -> tuple[torch.Tensor, torch.Tensor]:
-        cls_token = jit_model_prediction_sample[ViTTokenNames.CLS_TOKEN_NAME]
-        patch_tokens = jit_model_prediction_sample[ViTTokenNames.PATCH_TOKEN_NAME]
-        return cls_token, patch_tokens
+    def forward(self, cls_token: torch.Tensor, patch_tokens: Optional[torch.Tensor]) -> torch.Tensor:
+        """
+        Returns vision transformer embedding.
 
-    def embed_cls_only(self, jit_model_prediction_sample: dict[str, Any]) -> torch.Tensor:
-        cls_token, _ = self.get_output_tokens(jit_model_prediction_sample)
+        Parameters
+        ----------
+        cls_token :
+            class token of shape (B, feature_dim)
+        patch_tokens :
+            patch tokens of (B, num_patches, feature_dim)
+
+        Returns
+        -------
+        output: torch.Tensor
+            Embedding of shape [B, feature_dim]
+
+        """
+        output = self.embed_fn(cls_token, patch_tokens)
+        return output
+
+    @staticmethod
+    def embed_cls_only(cls_token: torch.Tensor, patch_tokens: Optional[torch.Tensor]) -> torch.Tensor:
         return cls_token
 
-    def embed_patch_only(self, jit_model_prediction_sample: dict[str, Any]) -> torch.Tensor:
-        _, patch_tokens = self.get_output_tokens(jit_model_prediction_sample)
-        return patch_tokens
+    @staticmethod
+    def embed_patch_only(cls_token: torch.Tensor, patch_tokens: Optional[torch.Tensor]) -> torch.Tensor:
+        if patch_tokens is None:
+            raise ValueError("patch_tokens cannot be None for this embedding mode")
+        return patch_tokens.mean(dim=1)
 
-    def embed_mean(self, jit_model_prediction_sample: dict[str, Any]) -> torch.Tensor:
-        cls_token, patch_tokens = self.get_output_tokens(jit_model_prediction_sample)
+    @staticmethod
+    def embed_mean(cls_token: torch.Tensor, patch_tokens: Optional[torch.Tensor]) -> torch.Tensor:
+        if patch_tokens is None:
+            raise ValueError("patch_tokens cannot be None for this embedding mode")
         cls_token = cls_token.unsqueeze(1)
         tokens = torch.cat([cls_token, patch_tokens], dim=1)
         output = tokens.mean(dim=1)
         return output
 
-    def embed_concat_mean(self, jit_model_prediction_sample: dict[str, Any]) -> torch.Tensor:
-        cls_token, patch_tokens = self.get_output_tokens(jit_model_prediction_sample)
+    @staticmethod
+    def embed_concat_mean(cls_token: torch.Tensor, patch_tokens: Optional[torch.Tensor]) -> torch.Tensor:
+        if patch_tokens is None:
+            raise ValueError("patch_tokens cannot be None for this embedding mode")
         cls_token = cls_token.unsqueeze(1)
         mean_patch_token = patch_tokens.mean(dim=1).unsqueeze(1)
         tokens = torch.cat([cls_token, mean_patch_token], dim=1)
         output = tokens.mean(dim=1)
         return output
 
-    def embed_concat(self, jit_model_prediction_sample: dict[str, Any]) -> torch.Tensor:
-        cls_token, patch_tokens = self.get_output_tokens(jit_model_prediction_sample)
+    @staticmethod
+    def embed_concat(cls_token: torch.Tensor, patch_tokens: Optional[torch.Tensor]) -> torch.Tensor:
+        if patch_tokens is None:
+            raise ValueError("patch_tokens cannot be None for this embedding mode")
         mean_patch_token = patch_tokens.mean(dim=1)
         output = torch.cat([cls_token, mean_patch_token], dim=1)
         return output
